requirements.txt
beautifulsoup4==4.9.3
pandas==1.2.1
pip==20.1.1
nltk==3.5
scikit-learn==0.23.2
goose3==3.1.7
newspaper3k==0.2.8
